+++ src/modules/Header/components/index.js
@@ -2,6 +2,7 @@ export {default as BoardSpecs} from './BoardSpecs';
 export {default as ContentButtonGroup} from './ContentButtonGroup';
 export {default as FullLogo} from './FullLogo';
 export {default as HeaderItem} from './HeaderItem';
+export {default as HeaderTitle} from './HeaderTitle';
 export {default as IconGroup} from './IconGroup';
-export {default as Title} from './Title';
+export {default as SlideDownBG} from './SlideDownBG';
 export {default as TitledIcon} from './TitledIcon';