+++ src/modules/Home/index.js
@@ -1,3 +1,0 @@
-import * as containers from "./containers";
-
-export default { containers };