+++ src/modules/Thread/components/Controller/index.js
@@ -1,0 +1,4 @@
+export {default as WatchController} from './WatchController'
+export {default as ArchiveController} from './ArchiveController'
+export {default as CommentController} from './CommentController'
+export {default as UpdateController} from './UpdateController'