+++ src/modules/index.js
@@ -1,8 +1,8 @@
-export {default as AlertContainer} from './AlertContainer'
-export {default as Board} from './Board'
-export {default as Header} from './Header'
-// export {default as HeaderPanels} from './HeaderPanels'
-// export {default as Navbar} from './Navbar'
-export {default as Settings} from './Settings'
-export {default as Thread} from './Thread'
-// export {default as ThreadControls} from './ThreadControls'
+export {default as App} from './App';
+export {default as Board} from './Board';
+export {default as ContentPanels} from './ContentPanels';
+export {default as Dashboard} from './Dashboard';
+export {default as Header} from './Header';
+export {default as MediaViewer} from './MediaViewer';
+export {default as Settings} from './Settings';
+export {default as Thread} from './Thread';