+++ src/server/routes/4chan/getArchive.js
@@ -2,7 +2,7 @@ import Axios from 'axios';
 
 import API from '-/config/api.4chan';
 import options from '-/config/proxy';
-import { writeObjToRoot } from '../../services/inspector';
+import { writeObjToRoot } from '../../utils/inspector';
 // TODO AS: Make a 4chan archive parser in server
 // import { parseArchive } from '../../parsers';