+++ src/views/index.jsx
@@ -1,16 +1,15 @@
 import React, { PropTypes } from 'react';
 
-import HomeView   from './HomeView';
+import DashboardView   from './DashboardView';
 import ContentView from './ContentView';
-import SettingsView from './SettingsView';
+import MediaView from './MediaView';
 
-
-const Views = ({ id, homeID, contentID, settingsID }) => {
+const Views = ({ id }) => {
     return (
         <div id={id} className="Views">
-            <HomeView id={homeID}/>
-            <ContentView id={contentID}/>
-            <SettingsView id={settingsID} />
+            <DashboardView />
+            <ContentView />
+            <MediaView />
         </div>
     );
 };
@@ -19,9 +18,6 @@ Views.displayName = 'Views';
 
 Views.propTypes = {
     id: PropTypes.string,
-    homeID: PropTypes.string,
-    contentID: PropTypes.string,
-    settingsID: PropTypes.string,
 };
 
 export default Views;